requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.13
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.5.0
django-reversion==3.0.9
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==4.6.3
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.3.1
Pillow==8.1.0
prettytable==2.0.0
pycparser==2.20
pyecharts==1.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
serializers==0.2.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
